Cache whole credential chain, not just final creds
